refactor: directory structure changed, now it doesn't use utils.js file, instead each function is in a correspoding file
